WI #1941 Review changes
