+++ TypeCobol.LanguageServer/Completion Factory/CompletionFactoryHelpers.cs
@@ -1,8 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using TypeCobol.Compiler;
 using TypeCobol.Compiler.CodeElements;
 using TypeCobol.Compiler.CodeModel;
 using TypeCobol.Compiler.Directives;
@@ -13,9 +11,8 @@ using TypeCobol.LanguageServer.VsCodeProtocol;
 
 namespace TypeCobol.LanguageServer
 {
-    public static class CompletionFactoryHelpers
+    internal static class CompletionFactoryHelpers
     {
-
         /// <summary>
         /// Help to resolve procedure name inside consumed tokens.
         /// Will return a string containing only a proc name or an entire qualified name for the procedure (depending on the given tokens)
@@ -32,6 +29,23 @@ namespace TypeCobol.LanguageServer
                                 && t.TokenType != TokenType.IN_OUT) // Take tokens until keyword found
                 .Where(t => t.TokenType == TokenType.UserDefinedWord)
                 .Select(t => t.Text));
+        }
+
+        /// <summary>
+        /// Get the matching node for the given CodeElement, returns null if not found.
+        /// </summary>
+        /// <param name="fileCompiler">Current file being compiled with its compilation results</param>
+        /// <param name="codeElement">Target CodeElement</param>
+        /// <returns>Corresponding Node instance, null if not found.</returns>
+        public static Node GetMatchingNode(FileCompiler fileCompiler, CodeElement codeElement)
+        {
+            var codeElementToNode = fileCompiler.CompilationResultsForProgram.ProgramClassDocumentSnapshot?.NodeCodeElementLinkers;
+            if (codeElementToNode != null && codeElementToNode.TryGetValue(codeElement, out var node))
+            {
+                return node;
+            }
+
+            return null;
         }
 
         public static IEnumerable<string> AggregateTokens(IEnumerable<Token> tokensToAggregate)
@@ -239,12 +253,6 @@ namespace TypeCobol.LanguageServer
             return new CompletionItem(variableDisplay) { insertText = variableArrangedQualifiedName, kind = CompletionItemKind.Variable };
         }
 
-        public static CompletionItem CreateCompletionItemForIndex(Compiler.CodeElements.SymbolInformation index, DataDefinition variable)
-        {
-            var display = string.Format("{0} (from {1})", index.Name, variable.Name);
-            return new CompletionItem(display) {insertText = index.Name, kind = CompletionItemKind.Variable};
-        }
-
         public static Case GetTextCase(string tokenText)
         {
             if (string.IsNullOrEmpty(tokenText)) return Case.Lower;
@@ -297,12 +305,14 @@ namespace TypeCobol.LanguageServer
             { ParameterDescription.PassingTypes.Output, "OUTPUT" },
             { ParameterDescription.PassingTypes.InOut, "IN-OUT" }
         };
+
         private static readonly Dictionary<ParameterDescription.PassingTypes, string> _CamelParams = new Dictionary<ParameterDescription.PassingTypes, string>()
         {
             { ParameterDescription.PassingTypes.Input, "Input"},
             { ParameterDescription.PassingTypes.Output, "Output" },
             { ParameterDescription.PassingTypes.InOut, "In-Out" }
         };
+
         private static readonly Dictionary<ParameterDescription.PassingTypes, string> _LowerParams = new Dictionary<ParameterDescription.PassingTypes, string>()
         {
             { ParameterDescription.PassingTypes.Input, "input"},
@@ -322,6 +332,7 @@ namespace TypeCobol.LanguageServer
                     return _LowerParams;
             }
         }
+
         public enum Case
         {
             Lower = 0,   // default value
